New parameter -ModuleAndVersionsToCheck Accepts hash table of specific modules w optional versions Other improvements

+++ PowerShell/Scripts/AutoModuleInstallAndUpdate/AutoModuleInstallAndUpdate.ps1
@@ -1,6 +1,6 @@
 
 <#PSScriptInfo
-.VERSION 0.8.0.5
+.VERSION 0.8.0.6
 .GUID 7f59e8ee-e3bd-4d72-88fe-24caf387e6f6
 .AUTHOR Brian Lalancette (@brianlala)
 .DESCRIPTION Automatically installs or updates PowerShell modules from the PowerShell Gallery
@@ -20,14 +20,12 @@
 <#
 .SYNOPSIS
     Automatically installs or updates PowerShell modules from the PowerShell Gallery
-.EXAMPLE
-    AutoModuleInstallAndUpdate.ps1 -Confirm:$false -UpdateExistingInstalledModules -KeepPriorModuleVersions
-.EXAMPLE
-    AutoModuleInstallAndUpdate.ps1 -ModulesToCheck Az,SharePointDSC -AllowPrerelease
 .PARAMETER Confirm
     Prompts prior to updating any existing modules or cleaning up files left over from old versions of modules.
 .PARAMETER ModulesToCheck
-    Allows you to include an optional comma-delimited list of modules to install or update.
+    Allows you to include an optional comma-delimited list (array) of modules to install or update.
+.PARAMETER ModulesAndVersionsToCheck
+    Accepts a hash table containing name-value pairs of modules and specific desired versions. To request the latest available version of a particular module, leave its value as an empty set of double quotes.
 .PARAMETER UpdateExistingInstalledModules
     Switch that indicates whether we should look for updates to any modules that are already installed on the current system. Disabled by default.
 .PARAMETER AllowPrerelease
@@ -36,8 +34,20 @@
     Switch that specifies we should attempt to update any modules that weren't originally installed from the PowerShell Gallery. Disabled by default.
 .PARAMETER KeepPriorModuleVersions
     Switch indicating that older versions of any updated/installed modules should be left in place. By default, any old module versions detected are uninstalled and removed from the file system.
+.EXAMPLE
+    AutoModuleInstallAndUpdate.ps1 -Confirm:$false -ModulesToCheck Az,MSOnline -Verbose
+    This will check and if necessary install/update both the Az and MSOnline modules to the latest published (non-prerelease) version, without prompting for confirmation.
+.EXAMPLE
+    AutoModuleInstallAndUpdate.ps1 -ModulesAndVersionsToCheck @{SharePointDSC = "3.4.0.0"; xCredSSP = "1.3.0.0"; xPSDesiredStateConfiguration = ""} -Verbose
+    This will check and if necessary install/update SharePointDSC version 3.4.0.0, xCredSSP version 1.3.0.0, and the latest version of xPSDesiredStateConfiguration, with verbose output.
+.EXAMPLE
+    AutoModuleInstallAndUpdate.ps1 -Confirm:$false -UpdateExistingInstalledModules -IncludeAnyManuallyInstalledModules
+    This command will effectively attempt a full unattended update of all PowerShell modules currently detected on the system, regardless of whether they were installed from the PowerShell Gallery.
 .LINK
     https://github.com/Microsoft/PFE
+
+.LINK
+    https://www.powershellgallery.com/packages/AutoModuleInstallAndUpdate
 .NOTES
     Created & maintained by Brian Lalancette (@brianlala), 2017-2019.
 #>
@@ -46,10 +56,11 @@
 param
 (
     [Parameter(Mandatory = $false)][bool]$Confirm = $true,
-    [Parameter(Mandatory = $false)][ValidateNotNullOrEmpty()][array]$ModulesToCheck,
-    [Parameter(Mandatory = $false)][switch]$UpdateExistingInstalledModules = $false,
+    [Parameter(Mandatory = $false, ParameterSetName = 'Latest')][ValidateNotNullOrEmpty()][array]$ModulesToCheck,
+    [Parameter(Mandatory = $false, ParameterSetName = 'Versioned')][ValidateNotNullOrEmpty()][hashtable]$ModulesAndVersionsToCheck,
+    [Parameter(Mandatory = $false, ParameterSetName = 'Latest')][switch]$UpdateExistingInstalledModules = $false,
     [Parameter(Mandatory = $false)][switch]$AllowPrerelease = $false,
-    [Parameter(Mandatory = $false)][switch]$IncludeAnyManuallyInstalledModules = $false,
+    [Parameter(Mandatory = $false, ParameterSetName = 'Latest')][switch]$IncludeAnyManuallyInstalledModules = $false,
     [Parameter(Mandatory = $false)][switch]$KeepPriorModuleVersions = $false
 )
 
@@ -72,16 +83,26 @@ Remove-Variable -Name modulesUpdated -Scope Global -Force -ErrorAction SilentlyC
 Remove-Variable -Name modulesUnchanged -Scope Global -Force -ErrorAction SilentlyContinue
 Remove-Variable -Name modulesRemoved -Scope Global -Force -ErrorAction SilentlyContinue
 
-if ($null -eq $ModulesToCheck)
+# If we didn't specify any modules to check on the command line, create an empty hash table that we can add items to later
+if ($null -eq $ModulesToCheck -and $null -eq $ModulesAndVersionsToCheck)
 {
-    [array]$ModulesToCheck = @()
+    [hashtable]$ModulesAndVersionsToCheck = @{}
+}
+else
+{
+    [hashtable]$ModulesAndVersionsToCheck = @{}
+    # Add each item in the $ModulesToCheck array to the $ModulesAndVersionsToCheck hash table, with blank version to specify the latest available
+    foreach ($ModuleName in $ModulesToCheck)
+    {
+        $ModulesAndVersionsToCheck[$ModuleName] = "" # Empty quotes for latest available version
+    }    
 }
 if ($UpdateExistingInstalledModules)
 {
     foreach ($existingInstalledModule in ((Get-Module -ListAvailable | Where-Object ModuleBase -like "$env:ProgramFiles\WindowsPowerShell\Modules\*")))
     {
         Write-Verbose -Message " - Adding existing installed module '$($existingInstalledModule.Name)' to list of modules to update."
-        $ModulesToCheck += $existingInstalledModule.Name
+        $ModulesAndVersionsToCheck[$($existingInstalledModule.Name)] = "" # Empty quotes for latest available version
     }
 }
 
@@ -123,7 +144,7 @@ try
         $AllowPrereleaseParameter = @{}
     }
     #endregion
-    if (($UpdateExistingInstalledModules) -or ($ModulesToCheck.Count -ge 1))
+    if (($UpdateExistingInstalledModules) -or ($ModulesAndVersionsToCheck.Count -ge 1))
     {
         Write-Output " - Checking for requested PowerShell modules..."
         # Because SkipPublisherCheck and AllowClobber parameters don't seem to be supported on Win2012R2 let's set whether the parameters are specified here
@@ -143,21 +164,32 @@ try
         {
             $skipPublisherCheckParameter = @{}
         }
-        Write-Verbose -Message " - `$ModulesToCheck: $($ModulesToCheck | Select-Object -Unique)"
-        foreach ($moduleToCheck in ($ModulesToCheck | Select-Object -Unique))
+        Write-Verbose -Message " - `$ModulesAndVersionsToCheck.Keys: $($ModulesAndVersionsToCheck.Keys | Sort-Object)"
+        foreach ($moduleToCheck in ($ModulesAndVersionsToCheck.Keys | Sort-Object))
         {
             Write-Host -ForegroundColor Cyan "  - Module: '$moduleToCheck'..."
             $Host.UI.RawUI.WindowTitle = "Checking '$moduleToCheck'..."
+            if ([string]::IsNullOrEmpty($ModulesAndVersionsToCheck.$moduleToCheck))
+            {
+                $requiredVersion = "latest available"
+                $requiredVersionParameter = @{}
+            }
+            else
+            {
+                $requiredVersion = $ModulesAndVersionsToCheck.$moduleToCheck
+                $requiredVersionParameter = @{RequiredVersion = $requiredVersion}
+            }
+
             [array]$installedModuleVersions = Get-Module -ListAvailable -FullyQualifiedName $moduleToCheck
-            if ($null -eq $installedModuleVersions -and (!(Get-InstalledModule -Name $moduleToCheck -ErrorAction SilentlyContinue)))
+            if ($null -eq $installedModuleVersions -and (!(Get-InstalledModule -Name $moduleToCheck @requiredVersionParameter -ErrorAction SilentlyContinue)))
             {
                 # Install requested module since it wasn't detected
-                $onlineModule = Find-Module -Name $moduleToCheck -ErrorAction SilentlyContinue
+                $onlineModule = Find-Module -Name $moduleToCheck -ErrorAction SilentlyContinue @requiredVersionParameter
                 if ($onlineModule)
                 {
                     Write-Host -ForegroundColor DarkYellow  "   - Module '$moduleToCheck' not present. Installing version $($onlineModule.Version)..." @noNewLineSwitch
                     $Host.UI.RawUI.WindowTitle = "Installing '$moduleToCheck'..."
-                    Install-Module -Name $moduleToCheck -ErrorAction Inquire -Force @allowClobberParameter @skipPublisherCheckParameter @verboseParameter
+                    Install-Module -Name $moduleToCheck -ErrorAction Inquire -Force @allowClobberParameter @skipPublisherCheckParameter @verboseParameter @requiredVersionParameter
                     if ($?)
                     {
                         Write-Host -ForegroundColor Green "  Done."
@@ -166,14 +198,14 @@ try
                 }
                 else
                 {
-                    Write-Host -ForegroundColor Yellow "   - Module '$moduleToCheck' not present, and was not found in the PowerShell Gallery for installation/update."
+                    Write-Host -ForegroundColor Yellow "   - Module '$moduleToCheck' $($requiredVersion -replace 'latest version','') not present, and was not found in the PowerShell Gallery for installation/update."
                     [array]$global:modulesUnchanged += $moduleToCheck
                 }
                 $Host.UI.RawUI.WindowTitle = $originalWindowTitle
             }
             else
             {
-                $installedModule = Get-InstalledModule -Name $moduleToCheck -ErrorAction SilentlyContinue
+                $installedModule = Get-InstalledModule -Name $moduleToCheck -ErrorAction SilentlyContinue @requiredVersionParameter
                 if ($installedModule)
                 {
                     # If we were successful in querying the module this way it was probably originally installed from the Gallery
@@ -182,12 +214,15 @@ try
                 else # Was probably pre-installed or installed manually
                 {
                     # Grab the newest version in case there are multiple
-                    $installedModule = ($installedModuleVersions | Sort-Object Version -Descending)[0]
+                    $installedModule = ($installedModuleVersions | Sort-Object -Property Version -Descending)[0]
                     $installedModuleWasFromGallery = $false
                 }
-                # Look for online updates to already-installed requested module
-                Write-Host "   - Module '$moduleToCheck' version $($installedModule.Version) is already installed. Looking for updates..." @noNewLineSwitch
-                $onlineModule = Find-Module -Name $moduleToCheck @AllowPrereleaseParameter -ErrorAction SilentlyContinue
+                if ($requiredVersion -ne $($installedModule.Version))
+                {
+                    # Look for online updates to or specific version of already-installed requested module
+                    Write-Host "   - Module '$moduleToCheck' version $($installedModule.Version) is already installed. Looking for $requiredVersion version..." @noNewLineSwitch
+                }
+                $onlineModule = Find-Module -Name $moduleToCheck @AllowPrereleaseParameter -ErrorAction SilentlyContinue @requiredVersionParameter
                 if ($null -eq $onlineModule)
                 {
                     Write-Host -ForegroundColor Yellow "Not found in the PowerShell Gallery!"
@@ -199,18 +234,18 @@ try
                     if ($installedModule.Version -eq $onlineModule.version)
                     {
                         # Online and local versions match; no action required
-                        Write-Host -ForegroundColor Gray "Already up-to-date ($($installedModule.Version))."
+                        Write-Host -ForegroundColor Gray "$moduleToCheck version $($installedModule.Version) already installed."
                         [array]$global:modulesUnchanged += $moduleToCheck
                     }
                     else
                     {
-                        Write-Host -ForegroundColor Magenta "Newer version $($onlineModule.Version) found!"
+                        Write-Host -ForegroundColor Magenta "$($requiredVersion -replace "latest available","Newer version $($onlineModule.Version)") found!"
                         if ($installedModule -and $installedModuleWasFromGallery)
                         {
                             # Update to newest online version using PowerShellGet
                             Write-Host "   - Updating module '$moduleToCheck'..." @noNewLineSwitch
                             $Host.UI.RawUI.WindowTitle = "Updating '$moduleToCheck'..."
-                            Update-Module -Name $moduleToCheck -Force @confirmParameter @verboseParameter @AllowPrereleaseParameter -ErrorAction Continue
+                            Update-Module -Name $moduleToCheck -Force @confirmParameter @verboseParameter @AllowPrereleaseParameter -ErrorAction Continue @requiredVersionParameter
                             if ($?)
                             {
                                 Write-Host -ForegroundColor Green "  Done."
@@ -239,6 +274,7 @@ try
                     }
                     # Now check if we have more than one version installed and remove prior versions unless we've specified otherwise
                     [array]$installedModuleVersions = Get-Module -ListAvailable -FullyQualifiedName $moduleToCheck
+                    $installedModuleVersions += Get-InstalledModule -Name $moduleToCheck -AllVersions -AllowPrerelease
                     if ($installedModuleVersions.Count -gt 1)
                     {
                         if ($KeepPriorModuleVersions)
@@ -248,12 +284,16 @@ try
                         else
                         {
                             # Remove all non-current module versions including ones that weren't put there via the PowerShell Gallery
-                            [array]$oldModules = $installedModuleVersions | Where-Object {$_.Version -ne $onlineModule.Version}
-                            foreach ($oldModule in $oldModules)
+                            [array]$oldModules = $installedModuleVersions | Where-Object {$_.Version -ne $onlineModule.Version} | Select-Object -Unique
+                            if ($oldModules.Count -ge 1)
+                            {
+                                Write-Verbose -Message "  - Older versions of module '$moduleToCheck' found ($($oldModules.Count))."
+                            }
+                            foreach ($oldModule in $oldModules | Where-Object {$oldModule.Name -ne "PackageManagement"}) # Don't want to risk accidentally blowing away the main module doing the updating...
                             {
                                 Write-Host "   - Uninstalling old version $($oldModule.Version) of '$($oldModule.Name)'..." @noNewLineSwitch
                                 $Host.UI.RawUI.WindowTitle = "Uninstalling old version of '$($oldModule.Name)'..."
-                                Uninstall-Module -Name $oldModule.Name -RequiredVersion $oldModule.Version -Force -ErrorAction SilentlyContinue @verboseParameter
+                                Uninstall-Module -Name $oldModule.Name -RequiredVersion $oldModule.Version -Force -ErrorAction SilentlyContinue @verboseParameter -AllowPrerelease
                                 if ($?) {Write-Host -ForegroundColor Green "  Done."}
                                 # Unload the old module in case it was automatically loaded in this console
                                 if (Get-Module -Name $oldModule.Name -ErrorAction SilentlyContinue)
@@ -262,25 +302,31 @@ try
                                     Remove-Module -Name $oldModule.Name -Force -ErrorAction Inquire @verboseParameter
                                     if ($?) {Write-Host -ForegroundColor Green "  Done."}
                                 }
-                                # Comment this out if it actually removes the module we've just installed/updated...
-                                Write-Host "   - Removing old module files from $($oldModule.ModuleBase)..." @noNewLineSwitch
-                                $Host.UI.RawUI.WindowTitle = "Cleaning up old version of '$($oldModule.Name)'..."
-                                Remove-Item -Path $oldModule.ModuleBase -Recurse -ErrorAction SilentlyContinue @confirmParameter @verboseParameter
-                                if ($?) {Write-Host -ForegroundColor Green "Done."}
-                                else
+                                if ($null -ne $oldModule.ModuleBase)
                                 {
-                                    Write-Output "."
+                                    if (Test-Path -Path $oldModule.ModuleBase -ErrorAction SilentlyContinue)
+                                    {
+                                        Write-Host "   - Removing old module files from $($oldModule.ModuleBase)..." @noNewLineSwitch
+                                        $Host.UI.RawUI.WindowTitle = "Cleaning up old version of '$($oldModule.Name)'..."
+                                        Remove-Item -Path $oldModule.ModuleBase -Recurse -ErrorAction SilentlyContinue @confirmParameter @verboseParameter
+                                        if ($?) {Write-Host -ForegroundColor Green "Done."}
+                                        else
+                                        {
+                                            Write-Output "."
+                                        }
+                                    }
+                                    # Check if the old module directory is still present for some reason
+                                    if (Test-Path -Path $oldModule.ModuleBase -ErrorAction SilentlyContinue)
+                                    {
+                                        Write-Warning -Message "Some or all of the path '$($oldModule.ModuleBase)' could not be removed - check permissions on this location."
+                                    }
+                                    else
+                                    {
+                                        Write-Verbose -Message "  - Successfully removed prior version $($oldModule.Version) of '$($oldModule.Name)'."
+                                        [array]$global:modulesRemoved += "$($oldModule.Name) version $($oldModule.Version)"
+                                    }
+                                    $Host.UI.RawUI.WindowTitle = $originalWindowTitle
                                 }
-                                if (Test-Path -Path $oldModule.ModuleBase -ErrorAction SilentlyContinue)
-                                {
-                                    Write-Warning -Message "Some or all of the path '$($oldModule.ModuleBase)' could not be removed - check permissions on this location."
-                                }
-                                else
-                                {
-                                    Write-Verbose -Message "  - Successfully removed prior version $($oldModule.Version) of '$($oldModule.Name)'."
-                                    [array]$global:modulesRemoved += "$($oldModule.Name) version $($oldModule.Version)"
-                                }
-                                $Host.UI.RawUI.WindowTitle = $originalWindowTitle
                             }
                         }
                     }
@@ -291,7 +337,7 @@ try
         if ($null -eq $installedModule)
         {
             # Module was not installed from the Gallery, so we look for it an alternate way
-            $installedModule = Get-Module -Name $moduleToCheck -ListAvailable | Sort-Object Version | Select-Object -Last 1
+            $installedModule = Get-Module -Name $moduleToCheck -ListAvailable | Sort-Object -Property Version | Select-Object -Last 1
         }
         Write-Host -ForegroundColor Cyan "  - Done checking/installing module '$moduleToCheck'."
         Write-Output "  --"
